Closes #1054: IOF fixes
Submitted by jsquyres, Reviewed by Rolf V., RM-Approved by tdd and Rich

Merge performed with:
{{
cd your_v1.2_checkout
svn merge -r 15462:HEAD https://svn.open-mpi.org/svn/ompi/tmp/iof-v1.2-merge .
}}

Which coresponded to these four changesets:

r14967:
A bunch of fixes from the /tmp/iof-fixes branch that fix up ''some''
(but not ''all'') of the problems that we have seen with iof:

 * Reading very large files via stdin redirected to orteun (Sun saw
  this)
 * Reading a little bit of a large file redirected to orterun's stdin
  and then either closing stdin or exiting the process

The Big Change was to make the proxy iof (the one running in non-HNP
orteds) send back a "I'm closing the stream" ACK back to the service
iof.  This tells the HNP that there will be nothing more coming from
that peer, and therefore the iof forward should be removed.

Many other minor cleanups/fixes, terminology changes, and
documentation additions are included in this commit as well.  However,
there are still some pretty big outstanding issues with IOF that are
not addressed either by #967 or this commit.  A few examples:

 * IOF was designed to allow multiple subscribers to a single stream.
  We're not entirely sure that this works (for one thing, there is
  nothing in the ORTE/OMPI code base that uses this functionality).
 * There are also resources leaked when processes/jobs exit (per
  Ralph's first comment on this ticket).
 * There is no feedback to close orterun's stdin when all subscribers
  to the corresponding stream have closed stdin.

r14981:
 * Forgot to update the iof null component when we revamped the IOF
  framework.  Updated pointers to match current definitions.
 * Trimmed some dead wood while I was at it:
  * No need for component close function that does nothing
  * Use BEGIN/END_C_DECLS
  * Use recent MCA param register function
  * Ditch MCA param orte_iof_debug (it wasn't used anywhere)
  * Use MCA param orte_iof_override properly in the code (i.e., look
    up the value once and use the cached value later)

r14984: Fix typo that accidentally always resulted in a "true" result.

r15013:
Fix a warning that Tim P. found. 

The warning was indicative of overly-complex code anyway. So I 
removed the "first" bool and simply use a sentinel value in seq_min to 
indicate that nothing has changed. Note that this is "correct enough" 
for the moment -- more fixes will come in this area with tickets #1049 
and/or #1051.