Closes #1122: Fix a segv on Cray systems during MPI_Finalize
Submitted by Brian, Reviewed by jjhursey, RM-Approved by tdd

At it's core we are making sure that we don't reference structures that
are invalid in the event engine. This is normally protected by the
OPAL_HAVE_WORKING_EVENTOPS flag, but it looks like this protection has
degraded a bit over the past few months. This commit brings things back.

r15796:
If we don't have event ops, we don't have a current_base, so
don't dereference the pointer (fixes a segfault Josh was seeing).