Closes #1181: Remove some warnings about rm from OSX Leopard builds
Submitted by Jeff, Reviewed by Brian, RM-Approved by tdd.

Patch applied which corresponds to the following two revisions.

r16669:
Change everywhere we do a "rm -f conftest*" to "rm -rf conftest" to
cover the case where a subdirectory is also built that needs to be
removed.

Note that there are other macros that we don't control (AC, AM, and/or
LT) that also exhibit this problem that we cannot fix.  :-\

r16672:
Per #1181, make our use of rm be consistent with the rest of AC/AM.
LT uses $RM, but AC/AM appear to use "rm ...".  So we'll go with
that.