Update the 1.3 branch to include several commits since the branch occurred:
r18746
Update the default mca params for LANL to block parallel file systems from being used for session directories

r18747, r18748, r18749
Update the PML selection logic to detect when a modex is required, and in those cases to only have rank=0 report its selected module. This is per the email thread on the devel list:

http://www.open-mpi.org/community/lists/devel/2008/06/4223.php

r18750
Add something to the .hgignore that may only exist on the Mac...but is annoying nonetheless.

r18751
First step: be kind to Jeff's disk space - let's abort without dumping core files all over the place

r18753
Restore the route initialization to the global server, if one is specified. This enables us to publish/lookup to a global server when using the direct routed module.

r18759
Reconnect the opal dss debug envar with the debug output