Closes #1512: fix BLCR SIGPIPE when checkpointing restarted apps
Submitted by jjhursey, Reviewed by timattox, RM-Approved by bbenton

r19601:
Fix a SIGPIPE that may occur when checkpointing a restarted
process. This was a result of calling system() in the BLCR CRS.
After inspection and testing it was determined that the operation
was no longer necessary. So the call was removed thus fixing the bug.