Closes #1551: Prettify map and allocation displays
Submitted by rhc, Reviewed by jsquyres, RM-approved by bbenton

r19608:
Modify the --display-map and --display-alloc per note on devel
list to reduce info for user understanding. 

Add --display-devel-map and --display-devel-alloc to display
all the detailed info we used to provide - it is only of
use/interest to developers anyway and confuses users.

r19655:
Prettify the user level display of allocation and map to make
it easier to see and understand.