Closes #1516: Solution for SM BTL problem
Submitted by bosilca, Reviewed by jsquyres, RM-approved by bbenton

r19561:
Fix the deadlock when we run out of resources on the BTLs.
Move the progress function from the BML into the PML.  The BTL
progress functions are now directly registered with the event library.

r19564:
Addendum to r19561: also remove a comment that is no longer true and 
some code that is commented out.