Closes #1576: orted-fail-to-start fixes
Submitted by rhc, Reviewed by jsquyres, RM-approved by bbenton

r19512: Update the orte wait function for Windows.

r19513: The correct function to close a socket. Thanks to George for noticing.

r19754:
Fix a problem in the plm "failed to start" code observed by Jeff.
When we are unable to launch to a specific node because it doesn't
exist or is down, the system would hang and/or segv. The reason
for the hang was that we were "firing" the orted exit trigger
prior to its timer event being defined - thus "locking" that
one-shot and preventing it from firing when we actually were
ready to use it. 

The segv was caused by the fact that we don't really know which
daemon failed to start (at least, in most cases), so we didn't
set a pointer to the aborted proc object. All we really wanted,
though, was to ensure that mpirun returned a non-zero exit status,
so the fix was to simply return the default error status.

r19755:
Cleanup the plm failed-to-start problem a little - ensure that
the event is always defined so we don't have to check when
trying to trigger it, thus avoiding potential race conditions.