Closes #1664: Update the MX driver
Submitted by bosilca, Reviewed by timattox, RM-approved by bbenton

r19496:
Fix the problem with the unexpected MX handler. The Solution so far is to 
mostly don't use this mechanism, as we have to be thread safe in order to 
really take full advantage of it. The unexpected handler is called by one 
of th MX threads, and we do not have control on the moment. In a non threaded 
case, this will completely destroy our recv requests queues, so the safest 
approach is to don't the unexpected handler.

r19910: Remove some unused code and a compiler warning.

r19979:
Move {{{__counter}}} outside the {{{#ifdef}}} section.
Cleanup the usage of {{{__counter}}}.